fix(daemon): make a committed outcome carry its whole shape, not just a marker
Six of the nine findings on this PR share one cause. `committedFailure` mints a
MARKER, but "committed" is a property of an OUTCOME, and an outcome has four
obligations: classify correctly, carry the result the caller needs, announce the
new state to subscribers, and survive the transport. The constructor enforces
only the first, so the other three are per-site conventions — ArchiveSession's
path follows them, and each newer path drops a different one. Nine symptoms, one
missing abstraction.

Transport (3732324438, 3732363049). Not a net/rpc limitation: ArchiveSession's
handler already answers OK with resp.Warning on a committed error, precisely
because net/rpc flattens a concrete error to rpc.ServerError and strips the
marker. KillSession, RestoreSession and RestoreArchived were never taught it, so
their markers could not reach any client. All four now share one named helper,
commitWarning, instead of copying the idiom — a new mutation handler cannot
half-adopt a function it has to call. Kill/Restore/RestoreArchived responses
gained the Warning field archive already had.

Results (3732277826, 3732277829). A committed error becomes a SUCCESSFUL
response, so returning "" and a zero InstanceData alongside it hands clients a
warning describing nothing — and DeleteProject appends that zero record to its
archived results. The remote path now returns its branch and the local one its
archivedPath plus projection.

The committed machinery moved to daemon/control_committed.go. control_server.go
was at ~995 lines, so any addition breached the 1000-line limit; splitting is
what CLAUDE.md asks for over grandfathering, and this concept is a coherent unit
— marker, constructor, API code, handler rule — that a new mutation path should
meet in one place.

Remaining findings (classification at archive.go:370/764, manager_sessions.go:206,
the inverse at :567, and the event at :764) are the same cause's other two faces
and are being addressed next; they are not covered by this commit.
